Move unused components to holding folder & change top20 folder name
